Refactoring

+++ internal/microservice/models/repository_models/answer.go
@@ -4,5 +4,5 @@ type Answer struct {
 	ID         uint32 `db:"id"`
 	QuestionID uint32 `db:"question_id"`
 	Login      string `db:"login"`
-	Mark       uint64 `db:"mark"`
+	Mark       uint32 `db:"mark"`
 }
